+++ tests/drivers/db/TestCase.php
@@ -10,6 +10,7 @@ namespace tests\drivers\db;
 use tests\app\PriorityJob;
 use tests\app\RetryJob;
 use tests\drivers\CliTestCase;
+use Yii;
 use yii\db\Query;
 
 /**
@@ -105,8 +106,48 @@ abstract class TestCase extends CliTestCase
         $this->assertEquals(0, $actual);
     }
 
+    public function testWaitingCount()
+    {
+        $this->getQueue()->push($this->createSimpleJob());
+
+        $this->assertEquals(1, $this->getQueue()->getStatisticsProvider()->getWaitingCount());
+    }
+
+    public function testDelayedCount()
+    {
+        $this->getQueue()->delay(5)->push($this->createSimpleJob());
+
+        $this->assertEquals(1, $this->getQueue()->getStatisticsProvider()->getDelayedCount());
+    }
+
+    public function testReservedCount()
+    {
+        $this->getQueue()->messageHandler = function () {
+            $this->assertEquals(1, $this->getQueue()->getStatisticsProvider()->getReservedCount());
+        };
+
+        $job = $this->createSimpleJob();
+        $this->getQueue()->push($job);
+        $this->getQueue()->run(false);
+    }
+
+    public function testDoneCount()
+    {
+        $this->getQueue()->messageHandler = function () {
+            return true;
+        };
+
+        $job = $this->createSimpleJob();
+        $this->getQueue()->push($job);
+        $this->getQueue()->run(false);
+
+        $this->assertEquals(1, $this->getQueue()->getStatisticsProvider()->getDoneCount());
+    }
+
+
     protected function tearDown()
     {
+        $this->getQueue()->messageHandler = null;
         $this->getQueue()->db->createCommand()
             ->delete($this->getQueue()->tableName)
             ->execute();